chat-service/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
fastapi-cors
python-multipart
httpx
requests    
beautifulsoup4
lxml
bs4
urllib3
requests-html
pydantic-settings
motor
pymongo
redis
aioredis
